Fixed login glitches
