+++ client/src/components/Home/desktop/LeftContent.vue
@@ -95,8 +95,8 @@
 </template>
 
 <script>
-import GraphQLService from "@/services/graphql.service";
 import FollowingList from "./FollowingList.vue";
+import LogoutFunction from "@/mixins/logout_function.js";
 
 export default {
     data() {
@@ -107,30 +107,8 @@ export default {
     components: {
         FollowingList
     },
+    mixins: [LogoutFunction],
     methods: {
-        logout() {
-            GraphQLService.logoutUser().then((response) => {
-                if (response.errors) {
-                    alert("Unable to logout");
-                    console.error(response.errors);
-                } else {
-                    // Flush local logged in user data
-
-                    this.$store.commit("refreshAccessToken", null);
-                    this.$store.commit("changeLoggedInState", false);
-                    this.$store.commit("changeUsername", null);
-
-                    // get rid of any localstorage cache
-                    localStorage.removeItem("profile_pic_link");
-
-                    // get rid of any user specific data in the store
-                    this.$store.dispatch("flush_user_data");
-                    this.$store.dispatch("flushPostCache");
-
-                    this.$store.dispatch("alertUser", { title: "Success", type: "success", msg: "Successfully logged out" });
-                }
-            });
-        },
         makeNewPost() {
             this.$parent.$parent.openPostMenu();
         },
