+++ server/package.json
@@ -38,7 +38,7 @@
         "morgan": "^1.10.0",
         "multer": "^1.4.2",
         "node-fetch": "^2.6.1",
-        "nodemailer": "^6.6.3",
+        "nodemailer": "^6.7.2",
         "nodemon": "^2.0.4",
         "sanitize-html": "^2.3.3"
     }
